Replace findhit-utils with lodash at runtime (fixes #36)
This lib is BSD, but had a runtime dependency on the GPL findhit-utils,
meaning any derivitive works need to be GPL.

findhit-util is still used as a dev dependency for testing the options
behaviour.

package.json "files" array has been added to not include the test code
in the resulting package so there's no GPL dependant code in the
resulting library.